src/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
django-environ==0.4.5
django_redis==4.11.0
psycopg2-binary==2.8.6
pytest==6.1.2
pytest-django==4.1.0
requests==2.25.0
